requirements.txt
pdfminer.six
pytesseract
pdf2image
Pillow
PyMuPDF
